// src/types/entry.ts
export type RdpResolution = "match_window" | "1920x1080" | "1280x720" | "1440x900" | "custom";

export interface SharedFolder {
  name: string;
  path: string;
  readOnly?: boolean;
}

export interface RdpEntryConfig {
  resolution: RdpResolution;
  customWidth?: number;
  customHeight?: number;
  colorDepth: 32 | 24 | 16 | 15;
  sound: "local" | "remote" | "none";
  quality: "best" | "good" | "low";
  clipboard: boolean;
  enableNla: boolean;
  hostname?: string;
  sharedFolders: SharedFolder[];
  enableHighDpi?: boolean;
}
// ...
export type RdpGlobalResolution = Exclude<RdpResolution, "custom">;

export interface RdpGlobalDefaults {
  resolution: RdpGlobalResolution;
  colorDepth: 32 | 24 | 16 | 15;
  sound: "local" | "remote" | "none";
  quality: "best" | "good" | "low";
  clipboard: boolean;
  enableNla: boolean;
  enableHighDpi: boolean;
  displayScale: number;
}